Fix imports and move remaining data files needed for session backward-compatibility testing
