+++ glue/plugins/tools/pv_slicer/qt/__init__.py
@@ -1,0 +1,1 @@
+from .pv_slicer import *
