+++ glue/plugins/tools/spectrum_tool/qt/__init__.py
@@ -1,0 +1,1 @@
+from .spectrum_tool import *
